pytest should run bottom-up now
